[GPU] Fix format dimension order portability

+++ src/plugins/intel_gpu/include/intel_gpu/runtime/format.hpp
@@ -341,7 +341,7 @@ struct format {
 
     static const std::vector<std::pair<size_t, int>> per_axis_block_size(format fmt);
 
-    static format find_format(const std::vector<uint64_t>& order,
+    static format find_format(const std::vector<size_t>& order,
                               const std::vector<std::pair<size_t, int>>& block_sizes,
                               bool is_weights = false,
                               bool is_grouped = false,
@@ -371,7 +371,7 @@ struct format {
     /// @brief Returns number of group dimensions.
     size_t group_num() const { return traits().group_num; }
     /// @brief Returns an order of dimensions.
-    const std::vector<uint64_t>& dims_order() const { return traits()._order; }
+    const std::vector<size_t>& dims_order() const { return traits()._order; }
     /// @brief Returns an order of dimensions in form of string.
     const std::string& order() const { return traits().order; }
     /// @brief Returns an internal orders of dimensions form of string.
